Commit H-bedd2c872c4dab1c

+++ auton_net/bot/database/developer_database.py
@@ -117,6 +117,6 @@ class GetDeveloperInformation:
             open_orders, integer
             rejected_orders, integer
             isAddressVerified, boolean
-            avg_completion_time, integer
+            avg_completion_time, float
             
         """
